+++ CHANGELOG.md
@@ -20,6 +20,8 @@
 * Added support for render html in `Publish box` with `publish_box` method.
 * Added `papi_get_entry_type_css_class` to get the css class that is added to body for a entry type.
 * Added support for autosaving fields.
+* Added support for group inside flexibles.
+* Added support for group inside repeaters.
 
 ### Changed